+++ CHANGELOG.md
@@ -9,14 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed
 ### Fixed
+- issue-65: Fixed footprints and previews disappearing from map when not intended and staying on too long.
 
 
-## [4.17.1]
+## [4.17.2]
 ### Added
 ### Changed
 ### Removed
 ### Fixed
 - issue-44: Fixed footprints and previews stay on map when not intended
+- issue-65: Fixed footprints and previews disappearing from map when not intended
 
 ## [4.17.0]
 ### Added
@@ -28,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - issue-48: Fixed collection resolution error handling
 - issue-49: Removed unused Docker and Jenkins folders/files from HiTIDE-UI
 - issue-52: Fixed missing thumbnail placeholder
+- issue-44: Fixed footprints and previews stay on map when not intended
 
 ## [4.16.2]
 ### Added